New /system-behavior api to control failure rate
Intead of relying on environment variables configured in the pods
to influence system behavior, it is not possible to configure some
behavior control from an API. The frontendservice maintains the
central set of behavior configurations and propagates these settings
during a request through an added `x-system-behavior` request header.

For now we only support configuring the checkoutservice behavior, but
the model is such that we could include future behavior controls as
well.

New supported apis:

GET /system-behavior
- returns the current system behavior settings

PATCH /system-behavior
- updates the provided settings for all subsequent requests